refactor: env get idx to return neg on not found

+++ includes/env.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 
 char *env_get(char **env, const char *const key);
-size_t env_get_idx(char **env, const char *const key);
+int env_get_idx(char **env, const char *const key);
 bool load_env_to_cache(t_ht *cache, char **env);
 bool env_var_name_validate(const char *var);
